% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{dend.heights}
\alias{dend.heights}
\title{all heights from a dendrogram}
\usage{
dend.heights(x)
}
\arguments{
\item{x}{a dendrogram}
}
\description{
all heights from a dendrogram
}
\keyword{internal}

