% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dend2hier}
\alias{dend2hier}
\title{Create a hierarchy from a dendrogram}
\usage{
dend2hier(x, height, newNames)
}
\arguments{
\item{x}{A dendrogram S3 R object.}

\item{height}{A vector of heights at which nodes are grouped.}

\item{newNames}{Names of the variabels which should be part of the 
hierarchy.}
}
\description{
A function which can be called from R. It is creating a 
hierarchy from a dendrogram.
}
\keyword{internal}

