% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{fast.anova}
\alias{fast.anova}
\title{Fast ANOVA}
\usage{
fast.anova(x, y, assign = NULL, family = gaussian(), test = c("LRT", "F"))
}
\arguments{
\item{x}{Design matrix of dimension \code{n * p}.}

\item{y}{Response vector of observations of length \code{n}.}

\item{assign}{Integer vector assigning columns to terms can be also given as 
\code{x} attribute in which case the argument is ignored. If an intercept exist it is 
expected to be the first column in \code{x} and it has to be specified by a '0' in tis 
vector. For details about assign see 
\code{\link[stats]{model.matrix}}.}

\item{family}{A description of the error distribution and link function to be used in 
the model. For GLMs this can be a character string naming a family function or the result 
of a call to a family function. (See \code{\link[stats]{family}} for details of family 
functions.)}

\item{test}{The name of the test either "LRT" (default) for likelihood ratio test or "F" 
for F test.}
}
\description{
A fast sequential analysis of variance (ANOVA). Mainly 
developed for internal use.
}
\examples{
y <- rnorm(n = 100)
x <- matrix(data = rnorm(1000), nrow = 100)
a <- 1:10
fast.anova(x = x, y = y, assign = a)

}
\seealso{
\code{\link[stats]{lm}}, \code{\link[stats]{anova}}, and 
\code{\link[stats]{aov}}.
}

