% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{fast.glmanova}
\alias{fast.glmanova}
\title{Fast GLM F test of LR test ANOVA}
\usage{
fast.glmanova(x, y, assign, family, test)
}
\arguments{
\item{x}{Design matrix of dimension \code{n * p}.}

\item{y}{Response vector of observations of length \code{n}.}

\item{assign}{Integer vector assigning columns to terms.}

\item{family}{A description of the error distribution and link function to be used in 
the model. For GLMs this must be the result of a call to a family function.}

\item{test}{The name of the test either "LRT" (default) for likelihood ratio  test or 
"F" for F test.}
}
\description{
A fast sequential analysis of variance (ANOVA). Mainly developed for 
internal use.
}
\keyword{internal}

