% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{fast.lmanova}
\alias{fast.lmanova}
\title{Fast LM F test ANOVA}
\usage{
fast.lmanova(x, y, assign)
}
\arguments{
\item{x}{Design matrix of dimension \code{n * p}.}

\item{y}{Response vector of observations of length \code{n}.}

\item{assign}{Integer vector assigning columns to terms.}
}
\description{
A fast sequential analysis of variance (ANOVA). Mainly developed for 
internal use.
}
\keyword{internal}

