% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/anova.R
\name{glm.anova}
\alias{glm.anova}
\title{Fast GLM F-test of LR-test ANOVA}
\usage{
glm.anova(x, y, assign, family, test)
}
\arguments{
\item{x}{Design matrix of dimension \code{n * p}.}

\item{y}{Response vector of observations of length \code{n}.}

\item{assign}{Integer vector assigning columns to terms can be also given as
\code{x} attribute in which case the argument is ignored. If an intercept
exist it is expected to be the first column in \code{x} and here it has to
specified by a '0'. For details about assign see
\code{\link[stats]{model.matrix}}.}

\item{family}{A description of the error distribution and link function to
be used in the model. For glm this must be the result of a call to a family
function. (See \code{\link[stats]{family}} for details of family functions.)}

\item{test}{The name of the test either "LRT" (default) for likelihood ratio
test or "F" for F-test}
}
\description{
A fast sequential analysis of variance (ANOVA). Mainly
developed for internal use.
}
\keyword{internal}

