% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{heightHierarchy}
\alias{heightHierarchy}
\title{All heights from a hierarchy}
\usage{
heightHierarchy(x)
}
\arguments{
\item{x}{a hierarchy}
}
\description{
All heights from a hierarchy
}
\keyword{internal}

