% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{hierarchy}
\alias{hierarchy}
\title{Hierachy indeces from desndrogram}
\usage{
hierarchy(x, height = NULL, names)
}
\arguments{
\item{x}{object of class 'dendrogram'.}

\item{height}{vector of cutting points.}

\item{names}{names in order.}
}
\description{
Hierachy indeces from desndrogram
}
\examples{
set.seed(123)
n <- 100
p <- 150
# x with correlated columns
sigmaChol <- chol(toeplitz((p:1/p)^3), pivot = TRUE)
sigmaChol <- sigmaChol[, order(attr(sigmaChol, "pivot"))]
x <- matrix(rnorm(n * p), nrow = n) \%*\% sigmaChol
x <- sweep(x, 2, rnorm(p, 10), "+")
colnames(x) <- paste0("x", 1:p)
# hierarchy
dend <- as.dendrogram(hclust(dist(t(x))))
hier <- hierarchy(dend)
}
\author{
Jonas Klasen
}

