% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hitPackage.R
\docType{package}
\name{hit-package}
\alias{HIT}
\alias{hit-package}
\title{Hierarchical Inference Testing}
\description{
Hierarchical inference testing (HIT) for (generalized) linear 
models with correlated covariates applicable to high-dimensional settings.
}
\author{
Jonas Klasen
}

