% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{names.hierarchy}
\alias{names.hierarchy}
\title{Names of Hierarchy}
\usage{
\method{names}{hierarchy}(x)
}
\arguments{
\item{x}{A \code{\link{as.hierarchy}}.}
}
\description{
Names of variables of an hierarchy.
}

