% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hit.R
\name{opt.penalty}
\alias{opt.penalty}
\title{Cross-validation of LASSO penenlty lambda}
\usage{
opt.penalty(x, y, family, nfolds, lambda.opt, alpha, penalty.factor, n.samp2,
  mc.cores, ...)
}
\arguments{
\item{x}{Design matrix, of dimension n x p.}

\item{y}{Vector of quantitative response variable.}

\item{family}{Distribution family of \code{y}.}

\item{nfolds}{Number of folds (default is 10). See}

\item{lambda.opt}{Criterion for optimum selection of cross-validated lasso. Either 
"lambda.1se" (default) or "lambda.min". See \code{\link[glmnet]{cv.glmnet}} for more 
details.}

\item{alpha}{A single value n the range of 0 to 1 for the elastic net mixing parameter.}

\item{penalty.factor}{See glmnet.}

\item{n.samp2}{Number of individuals in samp2 which is the max. for non zero 
coefficients.}

\item{mc.cores}{Number of cores for parallelising. Theoretical maximum is 'B'. For 
details see \code{\link[parallel]{mclapply}}.}

\item{...}{Additional agruments.}
}
\description{
Cross-validation of LASSO penenlty lambda.
}
\keyword{internal}

