% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hit.R
\name{p.hierarchy}
\alias{p.hierarchy}
\title{Significant hierarchy}
\usage{
p.hierarchy(x)
}
\arguments{
\item{x}{a hit object}
}
\description{
Significant hierarchy
}
\details{
makes a matrix of p-values for image(hit:::p.hierarchy(x))
}
\keyword{internal}

