% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hit.R
\name{p.matrix}
\alias{p.matrix}
\title{Significant hierarchy}
\usage{
p.matrix(x)
}
\arguments{
\item{x}{a hit object}
}
\description{
Significant hierarchy
}
\details{
makes a matrix of p-values for image(p.matrix(x))
}

