% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{reorder.hierarchy}
\alias{reorder.hierarchy}
\title{Reorder Hierarchy}
\usage{
\method{reorder}{hierarchy}(x, names, ...)
}
\arguments{
\item{x}{A \code{\link{as.hierarchy}}.}

\item{names}{Variable names in the order in which the indexes should be assigned to the 
variables.}

\item{...}{Further arguments passed to or from other methods (not used).}
}
\description{
Reorder indexes according to a vector of names.
}

