% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hit.R
\name{samp1.lambda.overall}
\alias{samp1.lambda.overall}
\title{Cross-validation of LASSO penenlty lambda}
\usage{
samp1.lambda.overall(x, y, family, alpha, lambda.opt, nfolds, penalty.factor,
  n.samp2, ...)
}
\arguments{
\item{x}{Design matrix, of dimension n x p.}

\item{y}{Vector of quantitative response variable.}

\item{family}{Distribution family of \code{y}.}

\item{alpha}{A single value n the range of 0 to 1 for the elastic net mixing parameter.}

\item{lambda.opt}{Criterion for optimum selection of cross-validated lasso. Either 
"lambda.1se" (default) or "lambda.min". See \code{\link[glmnet]{cv.glmnet}} for more 
details.}

\item{nfolds}{Number of folds (default is 10). See}

\item{penalty.factor}{See glmnet.}

\item{n.samp2}{Number of individuals in samp2 which is the max. for non zero 
coefficients.}

\item{...}{Additional agruments.}
}
\description{
Cross-validation of LASSO penenlty lambda.
}
\keyword{internal}

