% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hit.R
\name{samp2.sigHierarchy}
\alias{samp2.sigHierarchy}
\title{Variabel Testing}
\usage{
samp2.sigHierarchy(cIndex, level, upper.p, x, y, allSamp1.ids, allActSet.ids,
  x.nonTested, hierarchy, max.p.esti, B, gamma, cores)
}
\arguments{
\item{cIndex}{Index for cluster (mclapply index).}

\item{level}{Hierarchy level counter for parallelism}

\item{upper.p}{P value upper huerarchy level of the clusters variables.}

\item{x}{Design matrix, of dimension n x p.}

\item{y}{Vector of quantitative response variable.}

\item{allSamp1.ids}{List of subsampels.}

\item{allActSet.ids}{List of active sets.}

\item{x.nonTested}{Vector of indeces of non tested variabels.}

\item{hierarchy}{A hierarchy object}

\item{max.p.esti}{Maximum alpha level. All p-values above this value are set
to one. Small max.p.esti values reduce computing time.}

\item{B}{Number of sample-splits.}

\item{gamma}{Vector of gamma-values.}

\item{cores}{Number of cores for parallelising.}
}
\description{
ANOVA Testing, Multiplicity Adjustment, Aggregating and
Hierarchical Adjustment
}
\keyword{internal}

