% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hit.R
\name{samp2.testing}
\alias{samp2.testing}
\title{ANOVA testing, multiplicity adjustment, aggregating and hierarchical
adjustment}
\usage{
samp2.testing(clust.id, x, y, p.values, hierarchy, level, x.nonTested,
  allSamp1.ids, allActSet.ids, max.p.esti, B, gamma)
}
\arguments{
\item{clust.id}{index for cluster-ID (mclapply index).}

\item{x}{design matrix, of dimension n x p.}

\item{y}{vector of quantitative response variable.}

\item{p.values}{vector of p value per clusters.}

\item{hierarchy}{a hierarchy object}

\item{level}{hierarchy level}

\item{x.nonTested}{vector of indeces of non tested variabels.}

\item{allSamp1.ids}{list of subsampels.}

\item{allActSet.ids}{list of active sets.}

\item{max.p.esti}{maximum alpha level. All p-values above this value are set
to one. Small max.p.esti values reduce computing time.}

\item{B}{number of sample-splits.}

\item{gamma}{vector of gamma-values.}
}
\description{
ANOVA testing, multiplicity adjustment, aggregating and hierarchical
adjustment
}
\keyword{internal}

