% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hit.R
\name{summary.hit}
\alias{summary.hit}
\title{Summary of HIT}
\usage{
\method{summary}{hit}(object, alpha = 0.05, max.height, ...)
}
\arguments{
\item{object}{A \code{\link{hit}} object.}

\item{alpha}{A alpha significance threshold.}

\item{max.height}{max. Height to consider.}

\item{...}{Further arguments passed to or from other methods (not used).}
}
\description{
Significant clusters at alpha threshold.
}

