% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{heightDendrogram}
\alias{heightDendrogram}
\title{Heights of Dendrogram}
\usage{
heightDendrogram(x)
}
\arguments{
\item{x}{a \code{\link[stats]{dendrogram}}.}
}
\description{
All heights from a dendrogram.
}
\keyword{internal}

