% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{heightHierarchy}
\alias{heightHierarchy}
\title{Heights of Hierarchy}
\usage{
heightHierarchy(x)
}
\arguments{
\item{x}{a \code{\link{as.hierarchy}}.}
}
\description{
All heights from a hierarchy.
}
\keyword{internal}

