% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{hierarchy}
\alias{hierarchy}
\title{Hierachy indeces from desndrogram}
\usage{
hierarchy(dendrogram, min.dist = 0, max.dist = max(dist), h.levels = 100L,
  h.cut = NULL, mc.cores = 1L)
}
\arguments{
\item{dendrogram}{object of class 'dendrogram'.}

\item{min.dist}{min. distence to be tested.}

\item{max.dist}{max. distence to be tested.}

\item{h.levels}{number of hierarchy levels to be tested.}

\item{h.cut}{vector of cutting points. If it is specified min.dist,
max.dist, and h.levels are ignored.}

\item{mc.cores}{number of cores for parallelising,
see \code{\link{mclapply}}.}
}
\description{
Hierachy indeces from desndrogram
}
\examples{
set.seed(123)
n <- 100
p <- 150
# x with correlated columns
sigmaChol <- chol(toeplitz((p:1/p)^3), pivot = TRUE)
sigmaChol <- sigmaChol[, order(attr(sigmaChol, "pivot"))]
x <- matrix(rnorm(n * p), nrow = n) \%*\% sigmaChol
x <- sweep(x, 2, rnorm(p, 10), "+")
colnames(x) <- paste0("x", 1:p)
# hierarchy
dend <- as.dendrogram(hclust(dist(t(x))))
hier <- hierarchy(dend)
}
\author{
Jonas Klasen
}

