% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{hierarchy}
\alias{hierarchy}
\title{Hierachy indeces from desndrogram}
\usage{
hierarchy(x, height, max.height, names)
}
\arguments{
\item{x}{object of class 'dendrogram'.}

\item{height}{vector of heghts at which nodes are grouped.}

\item{max.height}{is the maximal heigt at which the testing starts.}

\item{names}{names in order.}
}
\description{
Hierachy indeces from desndrogram
}
\examples{
set.seed(123)
n <- 100
p <- 150
# x with correlated columns
corMat <- toeplitz(p:1/p)
corMatQ <- chol(corMat)
x <- matrix(rnorm(n * p), nrow = n) \%*\% corMatQ
colnames(x) <- paste0("x", 1:p)
# hierarchy
dend <- as.dendrogram(hclust(dist(t(x))))
hier <- hierarchy(dend, max.height = 20)
}
\author{
Jonas Klasen
}

