% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{hierarchy}
\alias{hierarchy}
\title{Clustering hierachy}
\usage{
hierarchy(x, height, max.height, names)
}
\arguments{
\item{x}{a \code{\link[stats]{dendrogram}}.}

\item{height}{vector of heights at which nodes are grouped.}

\item{max.height}{is the maximal height below the global node height which
is considered.}

\item{names}{variable names in the order in which the indexes shut be given
to the variables.}
}
\description{
Stores variable indexes of clustering hierarchies in a fast
accessible manner.
}
\examples{
set.seed(123)
n <- 100
p <- 150
# x with correlated columns
corMat <- toeplitz((p:1/p)^3)
corMatQ <- chol(corMat)
x <- matrix(rnorm(n * p), nrow = n) \%*\% corMatQ
colnames(x) <- paste0("x", 1:p)
# hierarchy
dend <- as.dendrogram(hclust(dist(t(x))))
hier <- hierarchy(dend, max.height = 20)
}
\author{
Jonas Klasen
}

