% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hit.R
\name{hit}
\alias{hit}
\title{Hierarchical inference testing}
\usage{
hit(x, y, hierarchy, B = 50, p.samp1 = 0.5, gamma = seq(0.05, 0.99, by =
  0.01), max.p.esti = 1, mc.cores = 1L, trace = FALSE, ...)
}
\arguments{
\item{x}{design matrix of dimension \code{n * p}, without intercept.
Variables not part of the dendrogram are added to the HO-model, see Details
below.}

\item{y}{quantitative response variable dimension \code{n}.}

\item{hierarchy}{object of class \code{\link{hierarchy}}. Must include all
variables of \code{x} which should be trested.}

\item{B}{number of sample-splits.}

\item{p.samp1}{fraction of data used for the LASSO. The ANOVA uses
\code{1 - p.samp1}.}

\item{gamma}{vector of gamma-values.}

\item{max.p.esti}{maximum alpha level. All p-values above this value are set
to one. Small \code{max.p.esti} values reduce computing time.}

\item{mc.cores}{number of cores for parallelising. Theoretical maximum is
'B'. For details see \code{\link[parallel]{mclapply}}.}

\item{trace}{if TRUE it prints current status of the program.}

\item{...}{additional arguments for \code{\link[glmnet]{cv.glmnet}}.}
}
\description{
Hierarchical inference testing for linear models with
high-dimensional and/or correlated covariates by repeated sample splitting.
}
\details{
The H0-model contains variables, with are not tested, like
experimental-design variables. These variables are not penalised in the
LASSO model selection and are always include in the reduced ANOVA model.
}
\examples{
set.seed(123)
n <- 100
p <- 150
# x with correlated columns
sigmaChol <- chol(toeplitz((p:1/p)^3), pivot = TRUE)
sigmaChol <- sigmaChol[, order(attr(sigmaChol, "pivot"))]
x <- matrix(rnorm(n * p), nrow = n) \%*\% sigmaChol
x <- sweep(x, 2, rnorm(p, 10), "+")
colnames(x) <- paste0("x", 1:p)
# y
y <- x[, c(3, 5, 73)] \%*\% c(2, 5, 3) + rnorm(n)
# hierarchy
dend <- as.dendrogram(hclust(dist(t(x))))
hier <- hierarchy(dend)
# HIT
out <- hit(x, y, hier)
summary(out)
}
\author{
Jonas Klasen
}
\references{
Mandozzi, J. and Buehlmann, P. (2013). \emph{Hierarchical testing in the
  high-dimensional setting with correlated variables.} To appear in the
  Journal of the American Statistical Association. Preprint arXiv:1312.5556
}

