% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{names.hierarchy}
\alias{names.hierarchy}
\title{Names of hierarchy}
\usage{
\method{names}{hierarchy}(x)
}
\arguments{
\item{x}{a \code{\link{hierarchy}}.}
}
\description{
Names of variables of an hierarchy.
}

