% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{names.hierarchy}
\alias{names.hierarchy}
\title{Leaf of the hierarchy
#}
\usage{
\method{names}{hierarchy}(x)
}
\arguments{
\item{x}{a hierarchy
#}

\item{x}{a hierarchy.}
}
\description{
#
}
\keyword{internal}

