% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{reorder.hierarchy}
\alias{reorder.hierarchy}
\title{Reorder Hierarchy}
\usage{
\method{reorder}{hierarchy}(x, names, ...)
}
\arguments{
\item{x}{a \code{\link{as.hierarchy}}.}

\item{names}{variable names in the order in which the indexes shut be given
to the variables.}

\item{...}{further arguments passed to or from other methods (not used).}
}
\description{
Reorder indexes according to a vector of names.
}

