% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{reorder.hierarchy}
\alias{reorder.hierarchy}
\title{reorder hierarchy according to names vector}
\usage{
\method{reorder}{hierarchy}(x, names, ...)
}
\arguments{
\item{x}{a hierarchy.}

\item{names}{names in new order.}

\item{...}{further arguments passed to or from other methods (not used).}
}
\description{
reorder hierarchy according to names vector
}

