% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hit.R
\name{samp1.lasso}
\alias{samp1.lasso}
\title{Variabel Screening}
\usage{
samp1.lasso(samp1, x, y, family, penalty.factor, lambda, n.samp2, ...)
}
\arguments{
\item{samp1}{List of index for subsample (mclapply index).}

\item{x}{Design matrix, of dimension n x p.}

\item{y}{Vector of quantitative response variable.}

\item{family}{Distribution family of \code{y}.}

\item{penalty.factor}{See glmnet.}

\item{lambda}{A vector of lambda values sorted from large to small where 
the smallest is the optimal value.}

\item{n.samp2}{Number of individuals in samp2 which is the max. 
for non zero coefficients.}

\item{...}{Additional agruments}
}
\description{
LASSO function of the HIT algorithem.
}
\keyword{internal}

