% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hit.R
\name{samp1.lasso}
\alias{samp1.lasso}
\title{LASSO screening}
\usage{
samp1.lasso(samp1, x, y, n.samp2, penalty.factor, ...)
}
\arguments{
\item{samp1}{list of index for subsample (mclapply index).}

\item{x}{design matrix, of dimension n x p.}

\item{y}{vector of quantitative response variable.}

\item{n.samp2}{number of individuals in samp2 which is the max.
for non zero coefficients.}

\item{penalty.factor}{see glmnet.}

\item{...}{aditional agruments}
}
\description{
LASSO screening
}
\keyword{internal}

