% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hit.R
\name{samp1.lasso}
\alias{samp1.lasso}
\title{Variabel Screening}
\usage{
samp1.lasso(samp1, x, y, n.samp2, lambda.opt, penalty.factor, nfolds, ...)
}
\arguments{
\item{samp1}{List of index for subsample (mclapply index).}

\item{x}{Design matrix, of dimension n x p.}

\item{y}{Vector of quantitative response variable.}

\item{n.samp2}{Number of individuals in samp2 which is the max.
for non zero coefficients.}

\item{lambda.opt}{Criterion for optimum selection of cross validated lasso.
Either 'lambda.min' (default) or 'lambda.1se'. See
\code{\link[glmnet]{cv.glmnet}} for more details.}

\item{penalty.factor}{See glmnet.}

\item{nfolds}{Number of folds (default is 10), see
\code{\link[glmnet]{cv.glmnet}} for more details.}

\item{...}{Additional agruments}
}
\description{
LASSO function of the HIT algorithem.
}
\keyword{internal}

