% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hit.R
\name{samp2.sigHierarchy}
\alias{samp2.sigHierarchy}
\title{Variabel Testing along the hierarchy}
\usage{
samp2.sigHierarchy(j, x, y, hierarchy, family, test, x.notest, B, allSamp1.ids,
  allActSet.ids, upper.p, max.p.esti, gamma, hl.count, cores)
}
\arguments{
\item{j}{Index for cluster (mclapply index).}

\item{x}{Design matrix, of dimension n x p.}

\item{y}{Vector of quantitative response variable.}

\item{hierarchy}{A hierarchy object.}

\item{family}{Distribution family of \code{y}.}

\item{test}{name of test.}

\item{x.notest}{Vector of indeces of non tested variabels.}

\item{B}{Number of sample-splits.}

\item{allSamp1.ids}{List of subsampels.}

\item{allActSet.ids}{List of active sets.}

\item{upper.p}{P value upper huerarchy level of the clusters variables.}

\item{max.p.esti}{Maximum alpha level. All p-values above this value are set to one. 
Small max.p.esti values reduce computing time.}

\item{gamma}{Vector of gamma-values.}

\item{hl.count}{Hierarchy level counter for parallelism.}

\item{cores}{Number of cores for parallelising.}
}
\description{
ANOVA Testing, Multiplicity Adjustment, Aggregating and Hierarchical 
Adjustment
}
\keyword{internal}

