% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hit.R
\name{samp2.sigHierarchy}
\alias{samp2.sigHierarchy}
\title{ANOVA testing, multiplicity adjustment, aggregating and hierarchical
adjustment}
\usage{
samp2.sigHierarchy(cIndex, level, upper.p, x, y, allSamp1.ids, allActSet.ids,
  x.nonTested, hierarchy, max.p.esti, B, gamma, cores)
}
\arguments{
\item{cIndex}{index for cluster (mclapply index).}

\item{level}{hierarchy level counter for parallelism}

\item{upper.p}{p value upper huerarchy level of the clusters variables.}

\item{x}{design matrix, of dimension n x p.}

\item{y}{vector of quantitative response variable.}

\item{allSamp1.ids}{list of subsampels.}

\item{allActSet.ids}{list of active sets.}

\item{x.nonTested}{vector of indeces of non tested variabels.}

\item{hierarchy}{a hierarchy object}

\item{max.p.esti}{maximum alpha level. All p-values above this value are set
to one. Small max.p.esti values reduce computing time.}

\item{B}{number of sample-splits.}

\item{gamma}{vector of gamma-values.}

\item{cores}{number of cores for parallelising.}
}
\description{
ANOVA testing, multiplicity adjustment, aggregating and hierarchical
adjustment
}
\keyword{internal}

