% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hit.R
\name{samp2.sigNode}
\alias{samp2.sigNode}
\title{ANOVA testing and multiplicity adjustment}
\usage{
samp2.sigNode(k, x, y, cluster, x.nonTested, allSamp1.ids, allActSet.ids)
}
\arguments{
\item{k}{index for subsample (mclapply index).}

\item{x}{design matrix, of dimension n x p.}

\item{y}{vector of quantitative response variable.}

\item{cluster}{clusters to be tested.}

\item{x.nonTested}{vector of indeces of non tested variabels.}

\item{allSamp1.ids}{list of subsampels.}

\item{allActSet.ids}{list of active sets.}
}
\description{
ANOVA testing and multiplicity adjustment
}
\keyword{internal}

