% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hit.R
\name{summary.hit}
\alias{summary.hit}
\title{Summary of HIT}
\usage{
\method{summary}{hit}(object, alpha = 0.05, max.height, ...)
}
\arguments{
\item{object}{a \code{\link{hit}} object.}

\item{alpha}{a alpha significance threshold.}

\item{max.height}{max. height to consider.}

\item{...}{further arguments passed to or from other methods (not used).}
}
\description{
Significant clusters at alpha threshold.
}

